fix some stiv warnings

+++ stiv.h
@@ -78,7 +78,7 @@ typedef union Number {
 } Number;
 
 #define UEBERZUG_CLEAR "{\"action\": \"remove\", \"identifier\": \"preview\"}\n"
-#define LENGTH(X) (sizeof(X)/sizeof(*X))
+#define LENGTH(X) (int64)(sizeof(X)/sizeof(*X))
 #define SNPRINTF(BUFFER, FORMAT, ...) \
     snprintf2(BUFFER, sizeof(BUFFER), FORMAT, __VA_ARGS__)
